app/base/base-config.[ch] use a gulong for the tile_cache_size.
2002-05-28  Sven Neumann  <sven@gimp.org>

	* app/base/base-config.[ch]
	* app/gui/user-install-dialog.c: use a gulong for the tile_cache_size.

	* app/base/tile-cache.c: cleanup, no changes.

	* app/config/gimpconfig-types.c (string_to_memsize): avoid overflows.

+++ app/base/base-config.h
@@ -26,7 +26,7 @@ struct _GimpBaseConfig
 {
   gchar    *temp_path;
   gchar    *swap_path;
-  guint     tile_cache_size;
+  gulong    tile_cache_size;
   gboolean  stingy_memory_use;
   gint      num_processors;
 };
